V2.01:
New features:
One-click-to-start switch in Expert screen
Can install itself to run on Windows bootup in background mode. Must be run as administrator and it will set up a Windows Scheduled task for bootup.

Bug Fixes:
Nasty bug stopped regions from registering (missing a http://)
An OAR could not be put in the first region
If Mysql failed to start a log message caused a crash.